Support creating conda environments from web addresses
Fixes https://github.com/conda/conda/issues/5562

This allows for passing web addresses through to conda env create,
like the following:

    conda env create -f https://domain.com/my-environment.yml

This uses the urllib module to download the file
and then passes things on as normal.

There are a few cases where we need to avoid explicit checks
about if this is a file or if we want to prepend the directory to the
filename.  In these cases we avoid these operations by checking if
the filename starts with "https://" or "http://"